// packages/destination-actions/src/destinations/facebook-custom-audiences/createAudience/generated-types.ts
// Generated file. DO NOT MODIFY IT BY HAND.

export interface Payload {
  /**
   * The account ID to use. Testing Only.
   */
  accountId: string
  /**
   * The access token to use. Testing Only.
   */
  accessToken: string
  /**
   * The name of the audience to create.
   */
  name: string
  /**
   * The description of the audience to create.
   */
  description?: string
  /**
   * The claim objective of the audience to create.
   */
  claimObjective?: string
  /**
   * The customer file source of the audience to create.
   */
  customerFileSource?: string
}